refactor: split widget layout to layouts
