+++ src/layouts/onboarding-layout/index.ts
@@ -1,0 +1,1 @@
+export * from "./view";
